Renamed all .h files to .hpp for assembler
