+++ assembler/inc/generators/project.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "cli/value.h"
+#include "cli/value.hpp"
 #include "psl/ustring.hpp"
 #include <filesystem>
