+++ assembler/inc/stdafx.hpp
