+++ assembler/src/cli_value.cpp
@@ -1,5 +1,5 @@
-﻿#include "cli_value.h"
-#include "stdafx.h"
+#include "cli_value.hpp"
+#include "stdafx.hpp"
 
 using namespace cli;
 uint64_t value_base::ID_GENERATOR {0u};
