+++ assembler/src/details/shader_cache.cpp
@@ -1,5 +1,5 @@
 #include "details/shader_cache.hpp"
-#include "stdafx.h"
+#include "stdafx.hpp"
 
 namespace assembler::details {
 auto shader_cache_t::get_transformed_content(entry_t const& entry) -> std::optional<psl::string> {
