+++ assembler/src/importers/audio.cpp
@@ -1,5 +1,5 @@
 #include "importers/audio.hpp"
-#include "stdafx.h"
+#include "stdafx.hpp"
 
 #include "core/meta/audio.hpp"
